Fix CI: satisfy Larastan level 7 (composer ci:check runs phpstan)
CI's ci:check script runs Pint, PHPStan (Larastan, level 7), then Pest;
PHPStan had never been run locally and reported 40 errors. All fixed at
the cause, no suppressions:

- Missing iterable/generic value types across the analysis core
  (MockDetector, front-ends, MannWhitney) and Eloquent relation generics
  on all three models; TestObservation documents the joined-in
  alias produced by DatasetQueries.
- PestFrontEnd now guards args with instanceof Node\Arg before touching
  ->value (VariadicPlaceholder has none) — uses(...A) spreads no longer
  hit an undefined property.
- Statistics call sites pass true lists (array_values) matching the
  list<> contracts; rank() honestly returns array<int,float> with float
  average ranks.
- BlameCommand: removed a dead ''-comparison on a non-empty-string.
- UserFactory::withTwoFactor() had an EMPTY BODY (returns null against
  a static return type) — latent starter-kit bug masked by the 2FA test
  being skipped; now returns a proper two-factor state.
- Livewire render() methods declare their View return type.

phpstan analyse: 0 errors. Tests 80 passed/1 skipped. Pint clean.

+++ app/Analysis/Statistics/MannWhitney.php
@@ -51,7 +51,12 @@ final class MannWhitney
         return ['u' => (float) $u, 'z' => $z, 'p' => max(0.0, min(1.0, $p))];
     }
 
-    /** Average-rank assignment with tie handling. @param list<int|float> $values */
+    /**
+     * Average-rank assignment with tie handling.
+     *
+     * @param  list<int|float>  $values
+     * @return array<int, float> rank per input index
+     */
     private static function rank(array $values): array
     {
         $indexed = [];
@@ -68,7 +73,7 @@ final class MannWhitney
             while ($j + 1 < $count && $indexed[$j + 1]['v'] === $indexed[$k]['v']) {
                 $j++;
             }
-            $avgRank = (($k + 1) + ($j + 1)) / 2; // ranks are 1-based
+            $avgRank = (float) ((($k + 1) + ($j + 1)) / 2); // ranks are 1-based
             for ($m = $k; $m <= $j; $m++) {
                 $ranks[$indexed[$m]['i']] = $avgRank;
             }
@@ -78,7 +83,10 @@ final class MannWhitney
         return $ranks;
     }
 
-    /** @param list<int|float> $values @return list<int> sizes of tie groups with size > 1 */
+    /**
+     * @param  list<int|float>  $values
+     * @return list<int> sizes of tie groups with size > 1
+     */
     private static function tieGroupSizes(array $values): array
     {
         $counts = [];
